requirements.txt
playwright==1.42.0
beautifulsoup4==4.12.3
aiohttp==3.9.3

fastapi
uvicorn
pydantic
google-genai
